+++ src/components/Card/index.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames/bind';
-import componentStyles from './style.css';
+import componentStyles from './index.module.css';
 import { COLOR } from '../../constants';
 
 const cx = classnames.bind(componentStyles);
 
 export const Card = ({ size, backgroundColor, boxShadow, children, ...props }) => {
-  const cardClass = cx('Card', { [`Card--${size}`]: size }, { 'Card--shadow': boxShadow });
+  const cardClass = cx('Card', `${size}`, { boxShadow });
 
   return (
     <div className={cardClass} style={{ backgroundColor }} {...props}>
