+++ src/components/Container/index.module.css
@@ -9,5 +9,6 @@
 
 .CardPasswordInput {
   display: flex;
-  flex-direction: row;
+  justify-content: flex-start;
+  margin-top: 0.25rem;
 }
