+++ src/components/Form/index.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames/bind';
-import componentStyles from './style.css';
+import componentStyles from './index.module.css';
 
 const cx = classnames.bind(componentStyles);
 
 export const Form = ({ className, children, ...props }) => {
-  const formClass = cx('Form', className);
+  const formClass = cx('Form', `${className}`);
 
   return (
     <form className={formClass} {...props}>
