+++ src/components/Modal/index.module.css
@@ -14,7 +14,7 @@
   justify-content: center;
 }
 
-.Modal--open {
+.isOpen {
   opacity: 1;
   visibility: visible;
 }
